renamed folder to pagetemplate
